Instance/EyeOfEternity: Fix warning

+++ src/server/scripts/Northrend/Nexus/EyeOfEternity/boss_malygos.cpp
@@ -1004,7 +1004,7 @@ public:
         void JustDied(Unit* /*killer*/)
         {
             // ugly hackfix for poke in the eye...
-            if (participants.size() < DUNGEON_MODE(ACHIEV_POKE_IN_THE_EYE_COUNT, ACHIEV_POKE_IN_THE_EYE_H_COUNT))
+            if (participants.size() < uint32(DUNGEON_MODE(ACHIEV_POKE_IN_THE_EYE_COUNT, ACHIEV_POKE_IN_THE_EYE_H_COUNT)))
                 if (AchievementEntry const* pAE = GetAchievementStore()->LookupEntry(DUNGEON_MODE(ACHIEV_POKE_IN_THE_EYE, ACHIEV_POKE_IN_THE_EYE_H)))
                     for (std::list<uint64>::const_iterator itr = participants.begin(); itr != participants.end(); ++itr)
                         if (Player* player = ObjectAccessor::FindPlayer((*itr)))
